Fix reset issue after QP failures
Close #19

When the QP fails, GlobalController::running is reset to false. If it is
not set back to true no command is sent to the simulated robot and so it
falls down. If one enable/disable step-by-step mode the flag is turned
on again which triggers the glitch seen in the issue.