Change the way data->qpos and data->qvel are initialized.
Whereas previously the joint ids was obtained implicitly by push_back in sequence, this commit uses mj_jnt_ids and mj_jnt_to_rjo to obtain the ids of the joint. This allows handling of cases where the number of joints differs between the mc_rtc robot and mjData. Such a case occurs in the simulation of a soft body (composite body).